Scrolling World

+++ inc/physics.h
@@ -5,15 +5,15 @@
 #include <time.h>
 
 struct Player {
-    float x;
-    float y;
+    int x;
+    int y;
     float vx;
     float vy;
     int width;
     int height;
 };
 
-Player* createPlayer(float x, float y, float vx, float vy, int width, int height);
+Player* createPlayer(int x, int y, float vx, float vy, int width, int height);
 void updatePhysics(Player* player, World* world, float dt);
 void destroyPhysics(Player* player);
